Extracting with_message logic to common module.
