+++ lib/matchers/validations/uniqueness_of.rb
@@ -2,6 +2,7 @@ module Mongoid
   module Matchers
     module Validations  
       class ValidateUniquenessOfMatcher < HaveValidationMatcher
+        include WithMessage
         def initialize(field)
           super(field, :uniqueness)
         end
@@ -19,11 +20,6 @@ module Mongoid
         
         def allow_blank?(allow_blank)
           @allow_blank = allow_blank
-          self
-        end
-
-        def with_message(message)
-          @expected_message = message
           self
         end
 
@@ -73,19 +69,6 @@ module Mongoid
             @positive_result_message << " with case insensitive values"
           else
             @negative_result_message << " without case insensitive values"
-            @result = false
-          end
-        end
-
-        def check_expected_message
-          actual_message = @validator.options[:message]
-          if actual_message.nil?
-            @negative_result_message << " with no custom message"
-            @result = false
-          elsif actual_message == @expected_message
-            @positive_result_message << " with custom message '#{@expected_message}'"
-          else
-            @negative_result_message << " got message '#{actual_message}'"
             @result = false
           end
         end
